+++ GameServerAdmin/GameServerAdmin/Models/Posts/AdminApi/DeletedPostDetailResponse.cs
@@ -2,12 +2,12 @@
 {
     public class DeletedPostDetailResponse
     {
-        public int PostId { get; set; }
+        public long PostId { get; set; }
         public string PostType { get; set; } = null!;
         public string Title { get; set; } = null!;
         public string Content { get; set; } = null!;
         public string AuthorType { get; set; } = null!;
-        public int AuthorId { get; set; }
+        public long AuthorId { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime? UpdatedAt { get; set; }
     }
